Removing block plots. Removing some markdown text.
The block plots aren't great. Would be nice to have
player trend line charts on single chart instead.

A lot of the text I'm removing is already in the README.
Also there is some that was OK in the notebook but
looks bad if you do a share w/o code. So cleaning that
up for sharing.