Fixes for new Cloud DB
DSX object storage has changed. These are fixes to the README
and notebook for this, as well as CloudantDB changes.

"Import StringIO object" is now "Import StreamingBody object" #24
The architecture diagram was missing Flow steps #25
Bluemix references are changed to IBM Cloud #26
The Cloudant credentials `url` is changed to `custom_url` #30
Adding Cloudant to DSX has changed, so docs are updated #31
When inserting a pandas dataframe, the data object has changed
 from `data_1` to `streaming_body_1` #32

Closes: #24
Closes: #25
Closes: #26
Closes: #30
Closes: #31
Closes: #32